Reduce DNS lookups for the statsd domain
The sendto socket call is very flexible but it currently
issue a DNS lookup for the statsd domain for every
metric in the for loop to push. This change ensures that
only one lookup is performed for each batch of metrics
to send. It is particularly useful when logster runs
every minute and the number of metrics to push is very high.