Implement proper locking logic for local filesystems
The current implementation of logsters cause logster to not to run again
if the lock file exists even if a process does not hold a lock on that.

This approach is fine for filesystems that does not have working locking
(e.g. some NFS, S3FS, etc.).

For native local filesystems failing only if the lock cannot be obtained
results in easier recovery from dead logsters.